+++ packages/core/build-cjs.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-cp .swcrc build && cd build && ../../../node_modules/.bin/swc ./esm -d ./cjs