+++ packages/system/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.55.0](https://github.com/Effect-TS/core/compare/@effect-ts/system@0.54.4...@effect-ts/system@0.55.0) (2022-02-24)
+
+
+### Bug Fixes
+
+* remove the missing `then` from Cause and ParSeq ([93afdb8](https://github.com/Effect-TS/core/commit/93afdb8b8bfaa1b0e19c64ad3ea002d2f2b414f6))
+
+
+
+
+
+## [0.54.4](https://github.com/Effect-TS/core/compare/@effect-ts/system@0.54.3...@effect-ts/system@0.54.4) (2022-02-23)
+
+**Note:** Version bump only for package @effect-ts/system
+
+
+
+
+
+## [0.54.3](https://github.com/Effect-TS/core/compare/@effect-ts/system@0.54.2...@effect-ts/system@0.54.3) (2022-02-23)
+
+**Note:** Version bump only for package @effect-ts/system
+
+
+
+
+
+## [0.54.2](https://github.com/Effect-TS/core/compare/@effect-ts/system@0.54.1...@effect-ts/system@0.54.2) (2022-02-23)
+
+**Note:** Version bump only for package @effect-ts/system
+
+
+
+
+
+## [0.54.1](https://github.com/Effect-TS/core/compare/@effect-ts/system@0.54.0...@effect-ts/system@0.54.1) (2022-02-22)
+
+**Note:** Version bump only for package @effect-ts/system
+
+
+
+
+
+# [0.54.0](https://github.com/Effect-TS/core/compare/@effect-ts/system@0.53.0...@effect-ts/system@0.54.0) (2022-02-22)
+
+
+### Bug Fixes
+
+* remove anything called `then` to avoid bundlers to think modules are thenable ([38ecdb6](https://github.com/Effect-TS/core/commit/38ecdb623f4331eb5976eaefe7ebb74301e8bf69))
+
+
+### BREAKING CHANGES
+
+* Cause.then is now combineSeq, Cause.both is now combinePar, Transducer.then is now andThen
+
+
+
+
+
+# [0.53.0](https://github.com/Effect-TS/core/compare/@effect-ts/system@0.52.4...@effect-ts/system@0.53.0) (2022-02-21)
+
+**Note:** Version bump only for package @effect-ts/system
+
+
+
+
+
 ## [0.52.4](https://github.com/Effect-TS/core/compare/@effect-ts/system@0.52.3...@effect-ts/system@0.52.4) (2022-02-17)
 
 **Note:** Version bump only for package @effect-ts/system